refactor: extract download token helpers
